+++ src/model/index.js
@@ -60,11 +60,9 @@ module.exports = function (options) {
   const [ _schemaName = 'public', _tableName ] = parser.separateSchema(_table)
 
   const _cleanable = _schema.cleanable
-  const _primaryKey = _schema.extensions.find(({ type }) => type === 'primaryKey')
+  const _primaryKey = R.path([ 'primaryKey', 0 ], _schema.extensions)
   const _forceCreate = R.isNil(_schema.force) ? force : _schema.force
-  const _hasConstraint = _schema.extensions.some(({ type }) => (
-    [ 'unique', 'primaryKey' ].includes(type)
-  ))
+  const _hasConstraint = _primaryKey || R.path([ 'unique', 0 ], _schema.extensions)
 
   const _seeds = new Seeds({
     table: _table,
@@ -191,7 +189,6 @@ module.exports = function (options) {
   }
 
   const _getDropExtensionQueries = (extensions, excludeNames) => R.pipe(
-    // must be drop
     R.values,
     R.unnest,
     R.filter(({ type, name }) => (
@@ -262,20 +259,24 @@ module.exports = function (options) {
     )
   )
 
-  const _getExtensionChangesOf = (dbExtensions, extensions) => {
+  const _getExtensionChangesOf = (dbExtensions, schemaExtensions) => {
     const adding = []
     const excludeDrop = []
-    if (utils.notEmpty(extensions)) {
-      extensions.forEach((extension) => {
-        const { type } = extension
-        const dbExtension = _findExtensionWhere(dbExtensions[type], extension)
-        if (dbExtension) {
-          excludeDrop.push(dbExtension.name)
-        } else {
-          adding.push(_addExtension(_table, extension))
-        }
-      })
-    }
+    R.forEachObjIndexed(
+      R.when(
+        utils.notEmpty,
+        R.forEach((extension) => {
+          const { type } = extension
+          const dbExtension = _findExtensionWhere(dbExtensions[type], extension)
+          if (dbExtension) {
+            excludeDrop.push(dbExtension.name)
+          } else {
+            adding.push(_addExtension(_table, extension))
+          }
+        }),
+      ),
+      schemaExtensions,
+    )
     return [
       _getDropExtensionQueries(dbExtensions, excludeDrop),
       adding,